Move serializers into separate module
