--- rdoasis/algorithms/serializers.py
@@ -1,5 +1,0 @@
-from rgd import utility
-
-from . import models
-
-utility.make_serializers(globals(), models)
